Move calculator and pollutant_type to package independent of etl
